Proper TOC datafile + major style changes
A .json file with a flat tree structure of a toc is now exported. The
structure is used to build the toc HTML and the tree nodes get links to
the DOM li elements of the toc for quick lookups. When loading a page
the index into the toc's is found and that let's us do all kinds of nice
stuff like traversing the hierarchy up from the current page to highlight
the toc or build the breadcrumb, and finding the previous/next page.