Use docfx html transform
Xrefs can't be resolved in the javascript code, so we need to let docfx
handle the transformation. docfx only runs the link transform on files
with the .html extension. I renamed the .html files to .html.tmp and
the .html.partial files to .html so they're properly processed by docfx.
I added a batch script to apply the renaming back from .html to
.html.partial and from .html.tmp to .html to the resulting files. Sucks
that this is necessary.